requirements.txt
Django==4.2.5
Pillow==10.0.1
django-debug-toolbar==4.2.0
setuptools==65.5.0
python-dotenv==1.0.0
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3
django-cors-headers==4.2.0
gunicorn==21.2.0
mysqlclient
django-storages
azure-storage-blob
azure-servicebus